Dialog and preference about renaming .pde to .ino files on save.
The new extension (.ino) is used by default for all new sketches (whether created with new or save as).  It's possible, however, to control the behavior on save.  The first time you save a sketch with a .pde file, you're prompted to rename it or cancel the save.  There's a preference that allow selecting whether or not .pde files are renamed on save.

http://code.google.com/p/arduino/issues/detail?id=644