Use pin 11 on a Mega for GSM soft serial
By default the GSM library uses digital pin 10 to communicate with the Mega but pin 10 is also used for the Ethernet shield's chip select. So if the shield is connected together with the Ethernet shield, we need to use a different interrupt pin for communication. Not all pins on the Mega and Mega 2560 support change interrupts. Only the following can be used for RX: 10, 11, 12, 13, 14, 15, 50, 51, 52, 53, A8 (62), A9 (63), A10 (64), A11 (65), A12 (66), A13 (67), A14 (68), A15 (69).  We decided to use pin 11 so we need to change the __RXPIN__ definition in the GSM3SoftSerial in the GSM library.